requirements.txt
python-telegram-bot==20.3
pymongo
dnspython
pydantic==2.7.0
